Move demangler type nodes out of GNU3
Share DemangledTypeNode across demangler implementations and add
MAX_DEMANGLE_NODE_LENGTH for bounded formatting.
